feat: double-precision coverage, rebuilt CompareSolarCoverage, safer clone
SolarCoverageSimulationResult: backing Dictionary<DateTime, float> →
Dictionary<DateTime, double>. Public API (Values, Coverage indexer,
GetCoverage, FromJsonObject/ToJsonObject) follows. float's ~7
significant digits were silently truncating fractional coverage values
in Grasshopper panels (0.999999 became "1"); double keeps the precision
all the way through. JSON serialization round-trips identically (JSON
numbers don't distinguish float from double).

Create.SolarCoverageSimulationResult: drop (float) cast on the
litArea / totalArea clamp — full double precision is preserved end-to-end
now.

SolarModel.GetSolarSimulationResults<T>: stop using reflection-based
Core.Query.Clone(x) and switch to direct typed copy constructors
(SolarCoverageSimulationResult and SolarFaceSimulationResult), matching
the pattern already used by SolarCoverageSimulationResults and
GetSolarFaceSimulationResults. Reflection clone fell back as a safety
net for unknown subtypes.

SAMAnalytical.CompareSolarCoverage (Grasshopper): rebuild for
AnalyticalModel A/B comparison with TAS-imported vs SAM-computed shade
data. Match faces by InternalPoint3D distance instead of centroid
(safer for non-convex panels). Align DateTimes by ceiling-rounding both
sides to the next whole hour and matching on (month, day, hour)
ignoring year — handles 30-minute offsets and different base years
between the two engines. Outputs (linkedFace3Ds_A/B, unmatched_A) now
emit Face3D via GooSAMGeometryParam so they preview directly in the
Rhino viewport and pipe cleanly into ISAMGeometry params.

+++ SAM_SolarCalculator/SAM.Geometry.SolarCalculator/Create/SolarCoverageSimulationResult.cs
@@ -41,7 +41,7 @@ namespace SAM.Geometry.SolarCalculator
                 return null;
             }
 
-            List<Tuple<DateTime, float>> coverage = new List<Tuple<DateTime, float>>();
+            List<Tuple<DateTime, double>> coverage = new List<Tuple<DateTime, double>>();
             foreach (Tuple<DateTime, Radiation, List<Face3D>> tuple in sunExposure)
             {
                 if (tuple == null)
@@ -63,7 +63,7 @@ namespace SAM.Geometry.SolarCalculator
                     }
                 }
 
-                float c = (float)Math.Max(0.0, Math.Min(1.0, litArea / totalArea));
+                double c = Math.Max(0.0, Math.Min(1.0, litArea / totalArea));
                 coverage.Add(Tuple.Create(tuple.Item1, c));
             }
 
@@ -100,7 +100,7 @@ namespace SAM.Geometry.SolarCalculator
                 return null;
             }
 
-            List<Tuple<DateTime, float>> coverage = new List<Tuple<DateTime, float>>(dateTimes.Count);
+            List<Tuple<DateTime, double>> coverage = new List<Tuple<DateTime, double>>(dateTimes.Count);
             foreach (DateTime dateTime in dateTimes)
             {
                 // Use GetSunExposureFace3Ds so coverage is computed from geometry regardless of
@@ -108,7 +108,7 @@ namespace SAM.Geometry.SolarCalculator
                 // Radiation is null even when lit polygons exist).
                 List<Face3D> litFaces = solarFaceSimulationResult.GetSunExposureFace3Ds(dateTime);
                 double litArea = litFaces == null ? 0 : litFaces.Sum(x => x?.GetArea() ?? 0);
-                float c = (float)Math.Max(0.0, Math.Min(1.0, litArea / totalArea));
+                double c = Math.Max(0.0, Math.Min(1.0, litArea / totalArea));
                 coverage.Add(Tuple.Create(dateTime, c));
             }
